refactor: 인디케이터 및 색상 관련 컴포저블 공통으로 이동 및 분리
